requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
python-dotenv
bcrypt